requirements.txt
asgiref==3.4.1
bidict==0.21.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.9
Django==4.0
django-cors-headers==3.10.1
djangorestframework==3.13.1
dnspython==2.1.0
eventlet==0.33.0
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.2
idna==3.3
Pillow==8.4.0
pycparser==2.21
python-engineio==4.3.0
pytz==2021.3
requests==2.26.0
six==1.16.0
sqlparse==0.4.2
tzdata==2021.5
urllib3==1.26.7
websocket-client==1.2.3
zope.event==4.5.0
zope.interface==5.4.0
